Peer review from Wenxin
For delete operation, it is generally common to return a 200 status code with a message of success. 204 is used when there is no content to send back in the response.
The function is catching errors in getRequestsCon, but it would be beneficial to have error handling in the other functions as well to make the application more robust.